+++ host/adblock/RELEASE_NOTES.md
@@ -1,5 +1,8 @@
 # Unreleased
 
+- Updated webext-ad-filtering-solution to 1.15.0
+- Attempt to fix issue when no subscriptions are enabled (due to corrupted data), restoring to default subscriptions (EXT-373)
+
 # 6.9.2 - 2024-09-25
 
 This release contains only minor updates and under-the-hood changes.